Pf1 split spellbooks (#115)
* PF1-Update spellbook name, add caster level check

I updated the logic for showing the spell book name and concentration check to use the same logic that displaying the spellbook name in PF1 uses. I also started implementing a caster level check much how the concentration check is added--this is not done! I was doing all of this in the in-browser IDE over my break at work. I still have to update the associated file that does the rolling to use the concentration check button. This is also fully untested since it's all just being edited in the browser. I'm hoping to get some time tonight to be able to finish fleshing it out.

* Added handler for pf1 caster level

* Fixes after being able to run the code

* Added casterLevel i18n

* Refactored each spellbook to their own tab