Update pf1-actions.js (#234)
* Update pf1-actions.js

Looking at the stack traces that was posted in the issue, this line is the problem. websterguy in the pf1 channel suggested changing this line to this edit, and it's worked for everyone that's made the change locally. I haven't tested it myself as there's a couple other issues keeping me from updating to the new version of PF1, but from everything I've seen this should sort it out.

Resolves #229

* Update scripts/actions/pf1/pf1-actions.js

Co-authored-by: JustNoon <45368472+JustNoon@users.noreply.github.com>

Co-authored-by: Drental <74384067+Drental@users.noreply.github.com>
Co-authored-by: JustNoon <45368472+JustNoon@users.noreply.github.com>